Update SoftWarmBeds_Patch.xml
I just added a feature to SWB to prevent it from changing the menu icons for the beds when it changes their graphics. This seems to help people not getting confused when choosing what beds to build. I figured your mod could use it too ;-)
You might want to review the resutling icon scaling.